to_json: rewrite to use the ujson module
to_json() now uses the `ujson` module instead of the `json` module for
improved performance.

to_json() now accepts kwargs list which will be passed as such to the
ujson.dumps() function.

to_json() now returns the resultin JSON only if stream is None, in which
case the resulting JSON is written to the stream and nothing is
returned.